requirements.txt
argparse==1.4.0
GitPython==3.1.37
python-dotenv==1.0.1
